pyproject.toml
[project]
name = "fishing-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyautogui>=0.9.54",
    "pyscreeze>=1.0.1",
]
